Added initial ConfigParser
